Modified : circleCI - npm run test; Added: phoneNumber & photoURL in signIn; Modified build;
