+++ .circleci/config.yml
@@ -28,7 +28,7 @@ jobs:
           pkg-manager: npm
       - run:
           name: Run tests
-          command: npm testß
+          command: npm run test
 
 workflows:
   # Below is the definition of your workflow.
