+++ src/Utils/Firebase/index.ts
@@ -1,0 +1,293 @@
+import { initializeApp } from "firebase/app";
+import {
+  getAuth,
+  signInWithRedirect,
+  signInWithPopup,
+  GoogleAuthProvider,
+  createUserWithEmailAndPassword,
+  signInWithEmailAndPassword,
+  signOut,
+  onAuthStateChanged,
+  UserInfo,
+  NextOrObserver,
+} from "firebase/auth";
+import {
+  getFirestore,
+  doc,
+  getDoc,
+  setDoc,
+  collection,
+  writeBatch,
+  QueryDocumentSnapshot,
+} from "firebase/firestore";
+import {
+  ObjectToAdd,
+  AdditionalInformation,
+} from "../../Types/Users";
+
+const firebaseConfig = {
+  apiKey: process.env["REACT_APP_FIREBASE_API_KEY"],
+  authDomain: process.env["REACT_APP_FIREBASE_AUTH_DOMAIN"],
+  projectId: process.env["REACT_APP_FIREBASE_PROJECT_ID"],
+  storageBucket: process.env["REACT_APP_FIREBASE_STORAGE_BUCKET"],
+  messagingSenderId: process.env["REACT_APP_FIREBASE_MESSAGING_SENDER_ID"],
+  appId: process.env["REACT_APP_FIREBASE_APP_ID"],
+  measurementId: process.env["REACT_APP_FIREBASE_MEASUREMENT_ID"],
+};
+
+export const firebaseApp = initializeApp(firebaseConfig);
+const googleProvider = new GoogleAuthProvider();
+googleProvider.setCustomParameters({
+  prompt: "select_account",
+});
+export const auth = getAuth();
+export const signInWithGooglePopup = () =>
+  signInWithPopup(auth, googleProvider);
+export const signInWithGoogleRedirect = () =>
+  signInWithRedirect(auth, googleProvider);
+
+export const db = getFirestore();
+export const addCollectionAndDocuments = async <T extends ObjectToAdd>(
+  collectionKey: string,
+  objectsToAdd: T[]
+) => {
+  const collectionRef = collection(db, collectionKey);
+  const batch = writeBatch(db);
+
+  objectsToAdd.forEach((object) => {
+    const docRef = doc(collectionRef, object.title.toLowerCase());
+    batch.set(docRef, object);
+  });
+
+  await batch.commit();
+  console.log("done");
+};
+
+export const createUserDocumentFromAuth = async (
+  userAuth: UserInfo,
+  additionalInformation = {} as AdditionalInformation
+): Promise<void | QueryDocumentSnapshot<UserInfo>> => {
+  if (!userAuth) return;
+
+  const userDocRef = doc(db, "users", userAuth.uid);
+
+  const userSnapshot = await getDoc(userDocRef);
+  // console.log(`userSnapshot => ..... ${userSnapshot},
+  //  userDocRef => ......${userDocRef}`, userSnapshot.exists(), userAuth, userSnapshot);
+  if (!userSnapshot.exists()) {
+    const { displayName, email, phoneNumber, photoURL } = userAuth;
+    const createdAt = new Date();
+
+    try {
+      await setDoc(userDocRef, {
+        displayName,
+        email,
+        createdAt,
+        photoURL,
+        phoneNumber,
+        ...additionalInformation,
+      });
+    } catch (error) {
+      console.log("error creating the user", error);
+    }
+  }
+
+  return userSnapshot as QueryDocumentSnapshot<UserInfo>;
+};
+
+export const createAuthUserWithEmailAndPassword = async (
+  email: string,
+  password: string
+) => {
+  if (!email || !password) return;
+
+  return await createUserWithEmailAndPassword(auth, email, password);
+};
+
+export const signInAuthUserWithEmailAndPassword = async (
+  email: string,
+  password: string
+) => {
+  if (!email || !password) return;
+
+  return await signInWithEmailAndPassword(auth, email, password);
+};
+
+export const signOutUser = async () => await signOut(auth);
+
+export const onAuthStateChangedListener = (callback: NextOrObserver<UserInfo>) =>
+  onAuthStateChanged(auth, callback);
+
+export const getCurrentUser = (): Promise<UserInfo | null> => {
+  return new Promise((resolve, reject) => {
+    const unsubscribe = onAuthStateChanged(
+      auth,
+      (userAuth) => {
+        unsubscribe();
+        resolve(userAuth);
+      },
+      reject
+    );
+  });
+};
+
+export default firebaseApp;
+
+// import { initializeApp } from 'firebase/app';
+// import {
+//   getAuth,
+//   signInWithRedirect,
+//   signInWithPopup,
+//   GoogleAuthProvider,
+//   createUserWithEmailAndPassword,
+//   signInWithEmailAndPassword,
+//   signOut,
+//   onAuthStateChanged,
+//   NextOrObserver,
+//   User,
+// } from 'firebase/auth';
+// import {
+//   getFirestore,
+//   doc,
+//   getDoc,
+//   setDoc,
+//   collection,
+//   writeBatch,
+//   query,
+//   getDocs,
+//   QueryDocumentSnapshot,
+// } from 'firebase/firestore';
+
+// const firebaseConfig = {
+//   apiKey: process.env.REACT_FIREBASE_API_KEY,
+//   authDomain: process.env.REACT_FIREBASE_AUTH_DOMAIN,
+//   projectId: process.env.REACT_FIREBASE_PROJECT_ID,
+//   storageBucket: process.env.REACT_FIREBASE_STORAGE_BUCKET,
+//   messagingSenderId: process.env.REACT_FIREBASE_MESSAGING_SENDER_ID,
+//   appId: process.env.REACT_FIREBASE_APP_ID,
+//   measurementId: process.env.REACT_FIREBASE_MEASUREMENT_ID
+// };
+
+// export const firebaseApp = initializeApp(firebaseConfig);
+// // export const firebaseAuth = firebaseApp.auth();
+
+// const googleProvider = new GoogleAuthProvider();
+
+// googleProvider.setCustomParameters({
+//   prompt: 'select_account',
+// });
+
+// export const auth = getAuth();
+// export const signInWithGooglePopup = () =>
+//   signInWithPopup(auth, googleProvider);
+// export const signInWithGoogleRedirect = () =>
+//   signInWithRedirect(auth, googleProvider);
+
+// export const db = getFirestore();
+
+// export type ObjectToAdd = {
+//   title : string ;
+// };
+
+// export const addCollectionAndDocuments = async <T extends ObjectToAdd>(
+//   collectionKey,
+//   objectsToAdd: T[]
+// ): Promise<void> => {
+//   const collectionRef = collection(db, collectionKey);
+//   const batch = writeBatch(db);
+
+//   objectsToAdd.forEach((object) => {
+//     const docRef = doc(collectionRef, object.title.toLowerCase());
+//     batch.set(docRef, object);
+//   });
+
+//   await batch.commit();
+//   console.log('done');
+// };
+
+// // export const getCategoriesAndDocuments = async (): Promise<Category[]> => {
+// //   const collectionRef = collection(db, 'categories');
+// //   const q = query(collectionRef);
+
+// //   const querySnapshot = await getDocs(q);
+// //   return querySnapshot.docs.map(
+// //     (docSnapshot) => docSnapshot.data() as Category
+// //   );
+// // };
+
+// export type AdditionalInformation = {
+//   displayName?: string;
+// };
+
+// export type UserData = {
+//   createdAt: Date;
+//   displayName: string;
+//   email: string;
+// };
+
+// export const createUserDocumentFromAuth = async (
+//   userAuth: User,
+//   additionalInformation = {} as AdditionalInformation
+// ): Promise<void | QueryDocumentSnapshot<UserData>> => {
+//   if (!userAuth) return;
+
+//   const userDocRef = doc(db, 'users', userAuth.uid);
+
+//   const userSnapshot = await getDoc(userDocRef);
+
+//   if (!userSnapshot.exists()) {
+//     const { displayName, email } = userAuth;
+//     const createdAt = new Date();
+
+//     try {
+//       await setDoc(userDocRef, {
+//         displayName,
+//         email,
+//         createdAt,
+//         ...additionalInformation,
+//       });
+//     } catch (error) {
+//       console.log('error creating the user', error);
+//     }
+//   }
+
+//   return userSnapshot as QueryDocumentSnapshot<UserData>;
+// };
+
+// export const createAuthUserWithEmailAndPassword = async (
+//   email: string,
+//   password: string
+// ) => {
+//   if (!email || !password) return;
+
+//   return await createUserWithEmailAndPassword(auth, email, password);
+// };
+
+// export const signInAuthUserWithEmailAndPassword = async (
+//   email: string,
+//   password: string
+// ) => {
+//   if (!email || !password) return;
+
+//   return await signInWithEmailAndPassword(auth, email, password);
+// };
+
+// export const signOutUser = async () => await signOut(auth);
+
+// export const onAuthStateChangedListener = (callback: NextOrObserver<User>) =>
+//   onAuthStateChanged(auth, callback);
+
+// export const getCurrentUser = (): Promise<User | null> => {
+//   return new Promise((resolve, reject) => {
+//     const unsubscribe = onAuthStateChanged(
+//       auth,
+//       (userAuth) => {
+//         unsubscribe();
+//         resolve(userAuth);
+//       },
+//       reject
+//     );
+//   });
+// };
+
+// export default firebaseApp;
