+++ src/index.tsx
@@ -20,7 +20,7 @@ root.render(
   </React.StrictMode>
 );
 
-ServiceWorker.register();
+ServiceWorker.unregister();
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
